ecommerce/requirements.txt
Django==4.2
djangorestframework==3.14.0
redis==5.0.0
mysqlclient>=2.0,<3.0